requirements.txt
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
loguru==0.7.3 ; python_version >= "3.11" and python_version < "4.0"
numpy==2.3.3 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.3.3 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.33.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic-settings==2.11.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.11.10 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.11" and python_version < "4.0"
typing-inspection==0.4.2 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.11" and python_version < "4.0"
win32-setctime==1.2.0 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
